Worked on linting errors and bugs in find_magic
Morgan helped me by showing me that I needed my linter on. Saw a lot of bugs dealing with spelling and formatting. Also found a few bugs in find_magic. Needed to use global on the exit_flag. Also fixed the log.info for when it says: now tracking .txt